timberproject/requirement.txt
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cffi==1.14.5
colorama==0.4.3
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-pdfkit==0.3.1
djangorestframework==3.12.4
entrypoints==0.3
filelock==3.0.12
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.4.2
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.3.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.1
pdfkit==0.6.0
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
Pygments==2.8.1
pylint==2.6.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
qrcode==6.1
qtconsole==5.0.3
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
sqlparse==0.4.1
terminado==0.9.3
testpath==0.4.4
toml==0.10.1
tornado==6.1
traitlets==5.0.5
virtualenv==20.0.30
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
